+++ app.js
@@ -31,6 +31,7 @@ var usuariosRoutes = require('./routes/usuarios');
 var loginRoutes = require('./routes/login');
 var imagenesRoutes = require('./routes/imagenes');
 var maniobraRoutes = require('./routes/maniobra');
+var maniobrasRoutes = require('./routes/maniobras');
 var operadorRoutes = require('./routes/operador');
 var camionRoutes = require('./routes/camion');
 var contenedorRoutes = require('./routes/contenedor');
@@ -108,6 +109,7 @@ app.use('/contenedor', contenedorRoutes);
 app.use('/camion', camionRoutes);
 app.use('/operador', operadorRoutes);
 app.use('/maniobra', maniobraRoutes);
+app.use('/maniobras', maniobrasRoutes);
 app.use('/solicitud', solicitudRoute);
 app.use('/img', imagenesRoutes);
 app.use('/dropzone', dropzoneRoutes);
